feat: refactor data service for callable functions and authentication
Refactors how the component makes calls to endpoints so that it utilized Firebase callable functions
when `useFirebase` is set to true. A side effect of this change is that authentication is now
required for all users. This is accomplished with anonymous auth via Firebase.

BREAKING CHANGE: This component now requires `firebase-project-id` and `firebase-token` in order to
work with Firebase.